requirements.txt
openai==1.51.0
python-dotenv==1.0.1
streamlit==1.39.0